+++ examples/logger-trait/src/server/init.server.lua
@@ -1,0 +1,17 @@
+--[==[
+
+	Copyright (C) Re-Sync - All Rights Reserved
+
+	[init.server.lua]:
+		Server side of the 'logger-trait' Lagoon example.
+
+	[Author(s)]:
+		Vyon - https://github.com/Vyon
+
+--]==]
+
+-- Modules:
+local MyService = require(script.Services.MyService)
+
+-- Main:
+MyService:Log("Debug/Info", "Hello, World!")