+++ examples/middleware/src/client/init.client.lua
@@ -1,0 +1,20 @@
+--[==[
+
+	Copyright (C) Re-Sync - All Rights Reserved
+
+	[init.client.lua]:
+		Client side of the 'net-trait' Lagoon example.
+
+	[Author(s)]:
+		Vyon - https://github.com/Vyon
+
+--]==]
+
+-- Services:
+local ReplicatedStorage = game:GetService("ReplicatedStorage")
+
+-- Modules:
+local Module = require(ReplicatedStorage.Shared.Module)
+
+Module:Init()
+Module.Greet()