+++ src/Types/init.lua
@@ -13,6 +13,7 @@
 -- Modules:
 local Generic = require(script.Generic)
 local NetType = require(script.NetType)
+local LoggerType = require(script.LoggerType)
 
 -- Types:
 -- Public:
@@ -23,7 +24,13 @@ export type Signal<T...> = Generic.Signal<T...>
 export type Connection = Generic.Connection
 
 -- Traits:
+--	Net:
 export type Net = NetType.Net
 export type MiddlewareEntry = NetType.MiddlewareEntry
+
+--	Logger:
+export type LogContext = LoggerType.LogContext
+export type LogOptions = LoggerType.LogOptions
+export type Logger = LoggerType.Logger
 
 return {}